\name{finallab4package-package}
\alias{finallab4package-package}
\alias{finallab4package}
\docType{package}
\title{
\packageTitle{finallab4package}
}
\description{
\packageDescription{finallab4package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{finallab4package}
\packageIndices{finallab4package}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{finallab4package}

Maintainer: \packageMaintainer{finallab4package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
